Match engine generation parameters with Manim output parameters
This allows precise matching between the desired FPS output (and its
inverse, the timestep) in Manim and the timestep used by the
engine.

This means that the engine calculates exactly the solution at each
Manim frame, and no interpolation is needed. This should unlock
further optimizations.

However, this also makes the exact output dependant on the FPS,
because the different engine timestep causes a different number of
updates per second to be created; each update will then use a separate
epsilon matrix. This is not considered to be a large issue.