Significantly boost performance of path state tracking
Experimental evidence shows that keeping track of all paths explicitly
(through a VMobject) adds a lot of overhead. Profiling suggests that
this is due to the various bookkeeping required to:

1) keep track of all positions;

2) continuously look at all segments of all paths, even those that are
   far in the past and not needed anymore.

This change keeps track of all points directly in a custom animation,
and makes the drawn object "jump" directly (via a `move_to`) to the
next coordinates it should be drawn at.

Ths speedup is significant. For shorter videos, where only about 2000
iterations were generated, the improvement is at least 4x. For longer
videos, the improvement is even better, because the bookkeeping
required by the previous code would make the number of iterations per
second drop progressively over time (it does not happen now).